initial port to 1.21.5
- update: inflation machine onRemove

+++ common/src/main/java/com/chefmooon/breezebounce/common/block/BreezeBounceBlock.java
@@ -38,7 +38,7 @@ public class BreezeBounceBlock extends Block implements SimpleBreezeBounceBlock 
     }
 
     @Override
-    public void fallOn(Level level, BlockState blockState, BlockPos blockPos, Entity entity, float f) {
+    public void fallOn(Level level, BlockState blockState, BlockPos blockPos, Entity entity, double f) {
         if (entity.isSuppressingBounce()) {
             if (f > DOUBLE_BOUNCE_ACTIVATION_THRESHOLD) tryDoubleBounceSpread(level, blockState, blockPos);
         } else {
