+++ common/src/main/java/com/chefmooon/breezebounce/common/block/BreezeBounceStairBlock.java
@@ -48,7 +48,7 @@ public class BreezeBounceStairBlock extends StairBlock implements SimpleBreezeBo
     }
 
     @Override
-    public void fallOn(Level level, BlockState blockState, BlockPos blockPos, Entity entity, float f) {
+    public void fallOn(Level level, BlockState blockState, BlockPos blockPos, Entity entity, double f) {
         if (entity.isSuppressingBounce()) {
             if (f > DOUBLE_BOUNCE_ACTIVATION_THRESHOLD) tryDoubleBounceSpread(level, blockState, blockPos);
         } else {
